🚚 Move stream arbitraries into /arbitrary
