+++ src/arbitrary/_internals/StreamArbitrary.ts
@@ -1,14 +1,12 @@
+import { NextArbitrary } from '../../check/arbitrary/definition/NextArbitrary';
+import { NextValue } from '../../check/arbitrary/definition/NextValue';
+import { cloneMethod } from '../../check/symbols';
 import { Random } from '../../random/generator/Random';
 import { Stream } from '../../stream/Stream';
 import { stringify } from '../../utils/stringify';
-import { cloneMethod } from '../symbols';
-import { Arbitrary } from './definition/Arbitrary';
-import { convertFromNext, convertToNext } from './definition/Converters';
-import { NextArbitrary } from './definition/NextArbitrary';
-import { NextValue } from './definition/NextValue';
 
 /** @internal */
-class StreamArbitrary<T> extends NextArbitrary<Stream<T>> {
+export class StreamArbitrary<T> extends NextArbitrary<Stream<T>> {
   constructor(readonly arb: NextArbitrary<T>) {
     super();
   }
@@ -42,19 +40,3 @@ class StreamArbitrary<T> extends NextArbitrary<Stream<T>> {
     return Stream.nil();
   }
 }
-
-/**
- * Produce an infinite stream of values
- *
- * WARNING: Requires Object.assign
- *
- * @param arb - Arbitrary used to generate the values
- *
- * @remarks Since 1.8.0
- * @public
- */
-function infiniteStream<T>(arb: Arbitrary<T>): Arbitrary<Stream<T>> {
-  return convertFromNext(new StreamArbitrary(convertToNext(arb)));
-}
-
-export { infiniteStream };
